+++ R/Listeners.R
@@ -118,7 +118,9 @@ Listeners <- R6::R6Class(
             stop('Error in executing action function of milestone <',
                  milestone$get_name(), '>. \n',
                  'Please set a breakpoint in its action function to debug it. \n',
-                 'The browser() function can be helpful for a step-by-step diagnosis. ')
+                 'The browser() function can be helpful for a step-by-step diagnosis. \n',
+                 'To fully replicate the issue in debugging, set seed = ',
+                 trial$get_output('seed'), ' in trial(...). ')
           }
         )
         milestone$set_dry_run(FALSE)