+++ R/computeCumulativeAlphaSpent.R
@@ -2,8 +2,6 @@
 ## Compute cumulative alpha spent on the critical values at information
 ## fractions.
 ##
-#' @importFrom stats as.formula pnorm
-#' @importFrom mvtnorm pmvnorm
 computeCumulativeAlphaSpent <- function(critical_values, information_fraction){
 
   corr <- outer(information_fraction, information_fraction,